improve commit test cases

+++ backend/src/main/java/ai/verta/modeldb/versioning/RepositoryDAORdbImpl.java
@@ -282,29 +282,33 @@ public class RepositoryDAORdbImpl implements RepositoryDAO {
       ListBranchesRequest.Response listBranchesResponse =
           listBranches(
               ListBranchesRequest.newBuilder().setRepositoryId(request.getRepositoryId()).build());
-      String deleteBranchesHQL =
-          "DELETE FROM "
-              + BranchEntity.class.getSimpleName()
-              + " br where br.id.repository_id = :repositoryId AND br.id.branch IN (:branches)";
-      Query deleteBranchQuery = session.createQuery(deleteBranchesHQL);
-      deleteBranchQuery.setParameter("repositoryId", repository.getId());
-      deleteBranchQuery.setParameterList("branches", listBranchesResponse.getBranchesList());
-      deleteBranchQuery.executeUpdate();
+      if (!listBranchesResponse.getBranchesList().isEmpty()) {
+        String deleteBranchesHQL =
+            "DELETE FROM "
+                + BranchEntity.class.getSimpleName()
+                + " br where br.id.repository_id = :repositoryId AND br.id.branch IN (:branches)";
+        Query deleteBranchQuery = session.createQuery(deleteBranchesHQL);
+        deleteBranchQuery.setParameter("repositoryId", repository.getId());
+        deleteBranchQuery.setParameterList("branches", listBranchesResponse.getBranchesList());
+        deleteBranchQuery.executeUpdate();
+      }
 
       CommitPaginationDTO commitPaginationDTO =
           commitDAO.fetchCommitEntityList(
               session, ListCommitsRequest.newBuilder().build(), repository.getId());
-      commitPaginationDTO.getCommitEntities().forEach(
-          commitEntity -> {
-            if (commitEntity.getRepository().contains(repository)) {
-              commitEntity.getRepository().remove(repository);
-              if (commitEntity.getRepository().isEmpty()) {
-                session.delete(commitEntity);
-              } else {
-                session.update(commitEntity);
-              }
-            }
-          });
+      commitPaginationDTO
+          .getCommitEntities()
+          .forEach(
+              commitEntity -> {
+                if (commitEntity.getRepository().contains(repository)) {
+                  commitEntity.getRepository().remove(repository);
+                  if (commitEntity.getRepository().isEmpty()) {
+                    session.delete(commitEntity);
+                  } else {
+                    session.update(commitEntity);
+                  }
+                }
+              });
 
       session.delete(repository);
       session.getTransaction().commit();
@@ -534,6 +538,24 @@ public class RepositoryDAORdbImpl implements RepositoryDAO {
       session.delete(branchEntity);
       session.getTransaction().commit();
       return DeleteBranchRequest.Response.newBuilder().build();
+    }
+  }
+
+  @Override
+  public void deleteBranchByCommit(Long repoId, String commitHash) {
+    try (Session session = ModelDBHibernateUtil.getSessionFactory().openSession()) {
+      session.beginTransaction();
+
+      StringBuilder deleteBranchesHQLBuilder =
+          new StringBuilder("DELETE FROM ")
+              .append(BranchEntity.class.getSimpleName())
+              .append(" br where br.id.repository_id = :repositoryId ")
+              .append(" AND br.commit_hash = :commitHash ");
+      Query deleteBranchQuery = session.createQuery(deleteBranchesHQLBuilder.toString());
+      deleteBranchQuery.setParameter("repositoryId", repoId);
+      deleteBranchQuery.setParameter("commitHash", commitHash);
+      deleteBranchQuery.executeUpdate();
+      session.getTransaction().commit();
     }
   }
